+++ src/include/ecoz2.h
@@ -32,7 +32,8 @@ int ecoz2_prd_show_file(
         int to
         );
 
-typedef void (*hmm_learn_callback_t)(char*, long double);
+// double to facilitate rust wrapper
+typedef void (*hmm_learn_callback_t)(char*, double);
 
 int ecoz2_hmm_learn(
         int N,
